Add files via upload
Add class resources to github. The more I mess with it, the more I like it. I can access it from work, I can store code on it as well as other file, and I can download and upload from anywhere. The markdown language is a bit confusing now, but I'm looking into it. I think I'll continue to use it after this class.